movies_api/requirements.txt
asgiref==3.4.1
Django==3.0.4
django-filter==2.4.0
djangorestframework==3.12.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
